refactor(auth): move Google OAuth routes to dedicated controller for separation of concerns
